+++ core/src/main/kotlin-2.1/org/jetbrains/ktfmt/util/CompatibilityUtils.kt
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-@file:Suppress("IncorrectPackageName", "PackageDirectoryMismatch")
-
 package org.jetbrains.ktfmt.util
 
 import org.jetbrains.kotlin.psi.KtContextReceiverList
@@ -24,10 +22,11 @@ import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes
 
 fun KtContextReceiverList.listToVisit(): List<KtElement> {
-  return contextParameters().ifEmpty { contextReceivers() }
+  return contextReceivers()
 }
 
-val CONTEXT_PARAMETER_LIST = KtStubElementTypes.CONTEXT_PARAMETER_LIST
+val CONTEXT_PARAMETER_LIST = KtStubElementTypes.CONTEXT_RECEIVER_LIST
 
+/** No val/var keyword in destructuring statements until Kotlin 2.3. */
 val KtDestructuringDeclarationEntry.ownValOrVarKeywordText: String?
-  get() = ownValOrVarKeyword?.text
+  get() = null