+++ core/src/test/kotlin/org/jetbrains/ktfmt/testutil/CaseConfig.kt
@@ -1,22 +1,36 @@
 package org.jetbrains.ktfmt.testutil
 
 import java.nio.file.Path
+import org.jetbrains.ktfmt.format.FileType
 import org.jetbrains.ktfmt.format.FormattingOptions
 import org.jetbrains.ktfmt.format.TrailingCommaManagementStrategy
 
 class Directive(val name: String, val configure: CaseConfig.Builder.(String) -> Unit)
 
-class CaseConfig(val options: FormattingOptions, val checkIdempotency: Boolean) {
+class CaseConfig(
+    val options: FormattingOptions,
+    val checkIdempotency: Boolean,
+    val fileType: FileType,
+) {
 
   class Builder(base: FormattingOptions) {
     val options: FormattingOptions.Builder = base.toBuilder()
     var checkIdempotency: Boolean = true
+    var fileType = FileType.REGULAR
 
-    fun build(): CaseConfig = CaseConfig(options.build(), checkIdempotency)
+    fun build(): CaseConfig = CaseConfig(options.build(), checkIdempotency, fileType)
   }
 
   companion object {
     private val DIRECTIVES: Map<String, Directive> = listOf(
+        Directive("FILE_TYPE") {
+          fileType =
+              when (it) {
+                "REGULAR" -> FileType.REGULAR
+                "SCRIPT" -> FileType.SCRIPT
+                else -> throw IllegalArgumentException("Unsupported file type: $it")
+              }
+        },
         Directive("MAX_WIDTH") { options.maxWidth(it.toInt()) },
         Directive("BLOCK_INDENT") { options.blockIndent(it.toInt()) },
         Directive("CONTINUATION_INDENT") { options.continuationIndent(it.toInt()) },
@@ -39,31 +53,40 @@ class CaseConfig(val options: FormattingOptions, val checkIdempotency: Boolean) 
         .associateBy { it.name }
 
     private val DIRECTIVE_REGEX = Regex("""^// ([A-Z][A-Z0-9_]+)(?: +(.*))?$""")
+    private val SHEBANG_REGEX = Regex("""^#!.*$""")
 
     fun parse(code: String, origin: Path? = null): ParsedDirectives {
-      val header = code.lines().takeWhile { it.startsWith("//") }
       var directivesEnded = false
-      val parsedDirectives =
-          header
-              .mapIndexedNotNull { lineNumber, line ->
-                val matchResult = DIRECTIVE_REGEX.matchEntire(line)
-                if (matchResult == null) {
-                  directivesEnded = true
-                  return@mapIndexedNotNull null
-                }
-                val (name, value) = matchResult.destructured
-                val directive =
-                    DIRECTIVES[name]
-                        ?: when {
-                          directivesEnded ->
-                              error(
-                                  "$origin:${lineNumber + 1}: directive '$name' should be listed first in the file",
-                              )
-                          else -> error("$origin:${lineNumber + 1}: unknown directive '$name'")
-                        }
-                directive to value
-              }
-              .toMap()
+      val parsedDirectives = buildMap {
+        val lines =
+            code.lineSequence().withIndex().let {
+              // directives come after shebang
+              if (it.firstOrNull()?.value?.matches(SHEBANG_REGEX) == true) {
+                put(DIRECTIVES["FILE_TYPE"]!!, "SCRIPT")
+                it.drop(1)
+              } else it
+            }
+        val header = lines.takeWhile { it.value.startsWith("//") }
+
+        for ((lineNumber, line) in header) {
+          val matchResult = DIRECTIVE_REGEX.matchEntire(line)
+          if (matchResult == null) {
+            directivesEnded = true
+            continue
+          }
+          val (name, value) = matchResult.destructured
+          val directive =
+              DIRECTIVES[name]
+                  ?: when {
+                    directivesEnded ->
+                        error(
+                            "$origin:${lineNumber + 1}: directive '$name' should be listed first in the file",
+                        )
+                    else -> error("$origin:${lineNumber + 1}: unknown directive '$name'")
+                  }
+          put(directive, value)
+        }
+      }
       return ParsedDirectives(parsedDirectives)
     }
   }